Split ActivityType collection resources
