+++ app/collections/activity_types/permissions.js
@@ -1,0 +1,34 @@
+import ActivityTypes from './';
+
+ActivityTypes.allow({
+  insert () {
+    // Get current user ID
+    const currentUserId = Meteor.userId();
+
+    // Chack if user is administrator
+    const userIsAdmin = Roles.userIsInRole(currentUserId, 'admin');
+
+    // Only Admin user can insert
+    return userIsAdmin;
+  },
+  remove () {
+    // Get current user ID
+    const currentUserId = Meteor.userId();
+
+    // Chack if user is administrator
+    const userIsAdmin = Roles.userIsInRole(currentUserId, 'admin');
+
+    // Only Admin user can remove
+    return userIsAdmin;
+  },
+  update () {
+    // Get current user ID
+    const currentUserId = Meteor.userId();
+
+    // Chack if user is administrator
+    const userIsAdmin = Roles.userIsInRole(currentUserId, 'admin');
+
+    // Only Admin user can update
+    return userIsAdmin;
+  }
+});
