+++ app/collections/activity_types/schema.js
@@ -1,0 +1,7 @@
+import SimpleSchema from 'simpl-schema';
+
+export default ActivityTypesSchema = new SimpleSchema({
+  name: {
+    type: String
+  }
+});
