+++ app/collections/activity_types/server/hooks.js
@@ -1,4 +1,5 @@
 import ActivityTypes from '../';
+import UserEventLog from '/collections/user_event_log/';
 
 ActivityTypes.after.insert(function (userId, activityType) {
   // Add event log
